backend/speed-chat-backend/src/requirements.txt
boto3==1.34.34
python-dateutil==2.8.2
